mcp-servers/calendar/pyproject.toml
[project]
name = "clawnix-mcp-calendar"
version = "0.1.0"
description = "ClawNix MCP server for Google Calendar"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
clawnix-mcp-calendar = "server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
